Move stack one directory up
