+++ src/agent/datapath_reducer.rs
@@ -6,7 +6,7 @@ use crate::emulation_core::mips::datapath::{DatapathState, Stage};
 use crate::emulation_core::mips::fp_registers::FpRegisters;
 use crate::emulation_core::mips::gp_registers::GpRegisters;
 use crate::emulation_core::mips::memory::Memory;
-use crate::emulation_core::mips::stack::Stack;
+use crate::emulation_core::stack::Stack;
 use gloo_console::log;
 use std::rc::Rc;
 use yew::Reducible;
