+++ src/agent/messages.rs
@@ -3,7 +3,7 @@ use crate::emulation_core::mips::datapath::DatapathState;
 use crate::emulation_core::mips::fp_registers::FpRegisters;
 use crate::emulation_core::mips::gp_registers::GpRegisters;
 use crate::emulation_core::mips::memory::Memory;
-use crate::emulation_core::mips::stack::Stack;
+use crate::emulation_core::stack::Stack;
 use crate::emulation_core::{architectures::AvailableDatapaths, mips::datapath::Stage};
 use serde::{Deserialize, Serialize};
