+++ src/emulation_core/mips.rs
@@ -10,4 +10,3 @@ pub mod fp_registers;
 pub mod gp_registers;
 pub mod instruction;
 pub mod memory;
-pub mod stack;
