+++ src/emulation_core/stack.rs
